server/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
fastapi==0.124.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
openai==2.11.0
pydantic==2.12.5
pydantic_core==2.41.5
PyPDF2==3.0.1
python-multipart==0.0.20
requests==2.32.5
sniffio==1.3.1
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.38.0
python-dotenv==1.1.0
packaging==25.0
pdf2image==1.17.0
pillow==12.0.0
PyPDF2==3.0.1
pytesseract==0.3.13
pytest==8.3.4
pytest-asyncio==0.24.0
slowapi==0.1.9
